Moved mapserver configuration to pkg.
